Added redis container test
Also updated .gitignore node_modules glob pattern for a more generic one
(hopefully it works for the next modules dir too and not just this one).

Also used the redis test as a review and
daily reminder that you need some one form
or another of callbacks for asynchronous programming in javascript.
Maybe try some other time the async keyword and train some more with the
generated flow to check if the lack of clean error management is too
painful without promises.