Added redis and postgres glue for main backend server
It worked so far.

Also reviewed the differences during printing of strings and json in the
console and in the express http response. If you try interpolating a
js object as if it was a string it won't work. If you use the express
res.json() method with the object as the single argument it will work.
No need for body-parser. Maybe it is included by default nowadays.

[]Next step should be including a worker that communicates with redis
through events and does simulated heavy processing in the background.