Dockerized node instances with docker-compose
Made the few adaptations that were left to make the app run with a single
docker-compose command. Both node/express instances are running with
nodemon and live changes restart them. It's extremely comfy to work that
way and exhilarating to see everything working even in this simple hello
world app.

[]NEXT: make postgres create a table and then register the numbers that
have been ordered, and finally retrieve them before making them be
recalculated.
[]TODO: create simplest frontend that can make an ajax call
[]TODO: create a nginx docker instance that proxies traffic to either
the frontend or the backend servers
[]TODO: exchange the simple frontend with a react frontend at some point
[]TODO: make everything work with kubernetes instead of relying on
docker-compose
[]TODO: and consider using one continuous integration tool like travis
or circle ci or even jenkins